+++ src/Markdig.Tests/MiscTests.cs
@@ -174,7 +174,7 @@ $$
         public void CanUseHttpsPrefixForWWWAutoLinks()
         {
             var pipeline = new MarkdownPipelineBuilder().UseAutoLinks().Build();
-            var httpsPipeline = new MarkdownPipelineBuilder().UseAutoLinks(new AutoLinkOptions() { UseHttpsPrefixForWWWLinks = true }).Build();
+            var httpsPipeline = new MarkdownPipelineBuilder().UseAutoLinks(new AutoLinkOptions() { UseHttpsForWWWLinks = true }).Build();
 
             TestParser.TestSpec("www.foo.bar", "<p><a href=\"http://www.foo.bar\">www.foo.bar</a></p>", pipeline);
             TestParser.TestSpec("www.foo.bar", "<p><a href=\"https://www.foo.bar\">www.foo.bar</a></p>", httpsPipeline);